Rename `RedisConnection` to `Connection`
This renames the `RedisConnection` module to `Connection`. This
is to remove the redundancy of saying "redis" twice.

Closes #3
