+++ lib/hella-redis.rb
@@ -1,5 +1,5 @@
-require "hella-redis/version"
-require "hella-redis/redis_connection"
+require 'hella-redis/version'
+require 'hella-redis/connection'
 
 module HellaRedis
 end
