+++ lib/hella-redis/connection.rb
@@ -1,0 +1,22 @@
+require 'connection_pool'
+require 'redis'
+require 'redis-namespace'
+
+module HellaRedis
+
+  module Connection
+
+    def self.new(config)
+      ::ConnectionPool.new(:timeout => config.timeout, :size => config.size) do
+        ::Redis::Namespace.new(config.redis_ns, {
+          :redis => ::Redis.connect({
+            :url    => config.url,
+            :driver => config.driver
+          })
+        })
+      end
+    end
+
+  end
+
+end
